账单管理添加批量删除功能；实现公告管理功能

+++ backend/book-manager/src/main/java/org/epsda/bookmanager/pojo/Notice.java
@@ -22,14 +22,14 @@ import java.time.LocalDateTime;
 @TableName("notice")
 public class Notice {
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     @NotNull
     @Length(max = 20, message = "公告标题最长不超过20位")
     private String title;
     @NotNull
     private String content;
     @NotNull
-    private Integer userId;
+    private Long userId;
     private Integer type;
     private Integer status;
     private LocalDateTime createTime;
